Cosmetic changes

+++ src/simlib/simaux.cpp
@@ -63,10 +63,10 @@ void gradient(cv::InputArray src, cv::OutputArray outX, cv::OutputArray outY) {
 }
 
 cv::Point2d imageCoordinate( cv::Point2d pt, cv::Mat im ) {
-   int ncols=im.cols, nrows=im.rows ;
+   double ncols=im.cols, nrows=im.rows ;
    return cv::Point2d( nrows/2 - pt.y, pt.x + ncols/2 ) ;
 }
 cv::Point2d pointCoordinate( cv::Point2d pt, cv::Mat im ) {
-   int ncols=im.cols, nrows=im.rows ;
+   double ncols=im.cols, nrows=im.rows ;
    return cv::Point2d( pt.y - ncols/2, nrows/2 - pt.x ) ;
 }
